requirements.txt
absl-py==1.0.0
aiocontextvars==0.2.2
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.1.0
analytics-python==1.4.0
antlr4-python3-runtime==4.8
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
backoff==1.10.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
contextlib2==21.6.0
contextvars==2.4
cryptography==36.0.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
datasets==1.17.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
dnspython==2.2.0
editdistance==0.6.0
entrypoints==0.4
fairseq==0.10.0
fastapi==0.74.1
ffmpy==0.3.0
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.6.0
google-auth-oauthlib==0.4.6
googletrans==3.0.0
gradio==2.8.4
grpcio==1.44.0
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.4.0
hydra-core==1.1.1
hyperframe==5.2.0
idna==2.10
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
langcodes==3.3.0
linkify-it-py==1.0.3
loguru==0.6.0
Markdown==3.3.6
markdown-it-py==2.0.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mdit-py-plugins==0.3.0
mdurl==0.1.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
monotonic==1.6
multidict==5.2.0
multiprocess==0.70.12.2
murmurhash==1.0.6
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.8
numpy==1.19.5
oauthlib==3.2.0
olefile==0.46
omegaconf==2.1.1
orjson==3.6.1
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.7.1
pathy==0.6.1
pickleshare==0.7.5
Pillow==8.4.0
pinecone-client==2.0.5
portalocker==2.4.0
preshed==3.0.6
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.8.2
pydeck==0.6.2
pydub==0.25.1
Pygments==2.11.2
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.7
pypiwin32==223
pyrsistent==0.18.0
python-dateutil==2.8.2
python-multipart==0.0.5
pyttsx==1.1
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
sacrebleu==2.0.0
sacremoses==0.0.47
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
sentry-sdk==1.5.4
six @ file:///tmp/build/80754af9/six_1623709665295/work
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
starlette==0.17.1
streamlit==1.5.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.10.2
torchaudio==0.10.2
torchtext==0.6.0
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
transformers==4.15.0
typer==0.4.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
tzdata==2021.5
tzlocal==4.1
uc-micro-py==1.0.1
urllib3==1.26.8
uvicorn==0.16.0
validators==0.18.2
wasabi==0.9.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
win32-setctime==1.1.0
wincertstore==0.2
xxhash==2.0.2
yarl==1.7.2
zipp==3.6.0
